--- bow/interpolation/type.go
@@ -1,23 +1,0 @@
-package interpolation
-
-import "errors"
-
-type Interpolation int
-
-const (
-	Linear = iota
-	StepPrevious
-)
-
-var (
-	ErrUnknownInterpolation = errors.New("unknown interpolation")
-)
-
-func Validate(interpolation Interpolation) error {
-	switch interpolation {
-	case Linear, StepPrevious:
-		return nil
-	default:
-		return ErrUnknownInterpolation
-	}
-}
