+++ bow/rolling/aggregation.go
@@ -5,7 +5,7 @@ import (
 	"fmt"
 
 	"github.com/metronlab/bow/bow"
-	"github.com/metronlab/bow/bow/transform"
+	"github.com/metronlab/bow/bow/rolling/aggregation/transform"
 )
 
 type ColumnAggregation interface {
