+++ bow/rolling/aggregation/arithmeticmean.go
