+++ bow/rolling/aggregation/arithmeticmean_test.go
