+++ bow/rolling/aggregation/count.go
