+++ bow/rolling/aggregation/count_test.go
