+++ bow/rolling/aggregation/difference.go
