+++ bow/rolling/aggregation/difference_test.go
