+++ bow/rolling/aggregation/integral.go
