+++ bow/rolling/aggregation/integral_test.go
@@ -4,7 +4,7 @@ import (
 	"testing"
 
 	"github.com/metronlab/bow/bow"
-	"github.com/metronlab/bow/bow/transform"
+	"github.com/metronlab/bow/bow/rolling/aggregation/transform"
 	"github.com/stretchr/testify/assert"
 )
