+++ bow/rolling/aggregation/minmax.go
