+++ bow/rolling/aggregation/mode.go
