+++ bow/rolling/aggregation/mode_test.go
