+++ bow/rolling/aggregation/sum.go
