+++ bow/rolling/aggregation/sum_test.go
