+++ bow/rolling/aggregation/transform/factor.go
