+++ bow/rolling/aggregation/transform/factor_test.go
