+++ bow/rolling/aggregation/transform/transform.go
