+++ bow/rolling/aggregation/weightedmean.go
