+++ bow/rolling/aggregation/weightedmean_test.go
