+++ bow/rolling/aggregation/wholeaggregation.go
