+++ bow/rolling/aggregation/wholeaggregation_test.go
