+++ bow/rolling/aggregation/windowstart.go
