+++ bow/rolling/fill/linear.go
