+++ bow/rolling/fill/windowstart.go
